+++ BivectorClock/BivectorClock.js
@@ -1,4 +1,4 @@
-import { FixedTime, WallClock } from "../sketchlib/animation/WallClock.js";
+import { FixedTime, AnalogClock } from "../sketchlib/animation/AnalogClock.js";
 import { Color } from "../sketchlib/Color.js";
 import { WIDTH, HEIGHT, SCREEN_CENTER } from "../sketchlib/dimensions.js";
 import { Direction } from "../sketchlib/pga2d/Direction.js";
@@ -94,7 +94,7 @@ const SCENE = group(
   style([HOUR_HAND, MINUTE_HAND, SECOND_HAND], STYLE_HANDS),
 );
 
-const WALL_CLOCK = new WallClock();
+const WALL_CLOCK = new AnalogClock();
 // uncomment to use a fixed time. It was chosen to keep the hands spread apart
 // at roughly 120 degrees which makes for a nice looking screenshot
 /*
